+++ interpolate.h
@@ -1,6 +1,6 @@
 #ifndef INTERPOLATE_H
 #define INTERPOLATE_H
 
-void interpolate(int i1, int d1, int i2, int d2, float values[]);
+void interpolate(float i1, float d1, float i2, float d2, float values[]);
  
 #endif
